pyproject.toml
[project]
name = "whatsapp-langgraph-agent-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "appdirs==1.4.4",
    "asttokens==3.0.0",
    "bs4>=0.0.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "decorator==5.2.0",
    "distro==1.9.0",
    "duckduckgo-search>=8.1.1",
    "executing==2.2.0",
    "fastapi==0.115.8",
    "google-search-results>=2.4.2",
    "groq==0.18.0",
    "gtts==2.5.4",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "ipython==8.32.0",
    "jedi==0.19.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-community>=0.3.18",
    "langchain-core==0.3.37",
    "langchain-groq==0.2.4",
    "langdetect>=1.0.9",
    "langgraph==0.2.74",
    "langgraph-checkpoint==2.0.16",
    "langgraph-sdk==0.1.53",
    "langsmith==0.3.10",
    "markdown-it-py==3.0.0",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "msgpack==1.1.0",
    "orjson==3.10.15",
    "packaging==24.2",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "prompt-toolkit==3.0.50",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pyee==11.1.1",
    "pygments==2.19.1",
    "pyppeteer==2.0.0",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "sarvamai>=0.1.5",
    "sniffio==1.3.1",
    "stack-data==0.6.3",
    "starlette==0.45.3",
    "tenacity==9.0.0",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.12.2",
    "urllib3==1.26.20",
    "uvicorn==0.34.0",
    "wcwidth==0.2.13",
    "websockets==10.4",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]
